✅ TST: move extend tests to sql_spec
